feat: add integration tests
